+++ schema/code.py
@@ -1,6 +1,7 @@
 from marshmallow.fields import Bool, DateTime as Date_Time, Nested, Str
 from schema.base import Base_Schema
 from utils.datetime import FORMAT
+from utils.schema import List_or_One
 
 __all__ = [
     'Code_Request_Schema'
@@ -13,5 +14,5 @@ class Code_Request_Schema(Base_Schema):
 
     id = Str(required=True, example='firewall',
                description='Code id.')
-    code = Str(required=True, many=True,
-                 description='Code source')
+    code = List_or_One(Str, required=True,
+                       description='Code source')