+++ CHANGELOG.md
@@ -15,3 +15,5 @@
 - Improved execution of action as daemon.
 - Accept list of strings for eBPF-Program code.
 - Stdout and stderr response format: plain, lines, and JSON.
+
+## 1.0.3